BAEL-8961 Fix surefire configs of spring-4, jasyptdemo, spring-boot logging and mvc, auth projects
- Test classes renaming as per standards
- Surefire config fixes, pom standardization
- Added packaging pom at spring-cloud-security module
