+++ spring-cloud/spring-cloud-security/auth-client/pom.xml
@@ -2,18 +2,15 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.baeldung</groupId>
     <artifactId>auth-client</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
     <name>auth-client</name>
-    <description>Demo project for Spring Boot</description>
-
+	<description>Spring Cloud Security APP Client Module</description>
+	
     <parent>
-        <artifactId>parent-boot-1</artifactId>
+        <artifactId>spring-cloud-security</artifactId>
         <groupId>com.baeldung</groupId>
-        <version>0.0.1-SNAPSHOT</version>
-        <relativePath>../../../parent-boot-1</relativePath>
+        <version>1.0.0-SNAPSHOT</version>
     </parent>
 
     <dependencies>
@@ -88,9 +85,6 @@
     </dependencyManagement>
 
     <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
         <js-cookie.version>2.1.0</js-cookie.version>
         <spring-cloud.version>Dalston.SR4</spring-cloud.version>
     </properties>
