+++ spring-cloud/spring-cloud-security/auth-client/src/test/java/com/example/springoath2/Springoath2ApplicationUnitTest.java
@@ -9,7 +9,7 @@ import com.baeldung.CloudSite;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = CloudSite.class)
-public class Springoath2ApplicationTests {
+public class Springoath2ApplicationUnitTest {
 
 	@Test
 	public void contextLoads() {
