+++ spring-cloud/spring-cloud-security/auth-resource/pom.xml
@@ -3,21 +3,18 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.baeldung</groupId>
     <artifactId>auth-resource</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>auth-resource</name>
-    <description>Demo project for Spring Boot</description>
+    <description>Spring Cloud Security APP Resource Module</description>
 
     <parent>
-        <artifactId>parent-boot-1</artifactId>
+        <artifactId>spring-cloud-security</artifactId>
         <groupId>com.baeldung</groupId>
-        <version>0.0.1-SNAPSHOT</version>
-        <relativePath>../../../parent-boot-1</relativePath>
+        <version>1.0.0-SNAPSHOT</version>
     </parent>
-
+    
     <dependencies>
         <dependency>
             <groupId>org.springframework.security.oauth</groupId>
@@ -60,9 +57,6 @@
     </build>
 
     <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
         <spring-cloud.version>Edgware.RELEASE</spring-cloud.version>
     </properties>
