+++ spring-cloud/spring-cloud-security/auth-server/pom.xml
@@ -2,15 +2,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.baeldung</groupId>
     <artifactId>auth-server</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-
+    <description>Spring Cloud Security APP Server Module</description>
+	
     <parent>
-        <artifactId>parent-boot-1</artifactId>
+        <artifactId>spring-cloud-security</artifactId>
         <groupId>com.baeldung</groupId>
-        <version>0.0.1-SNAPSHOT</version>
-        <relativePath>../../../parent-boot-1</relativePath>
+        <version>1.0.0-SNAPSHOT</version>
     </parent>
 
     <dependencies>
